app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
numpy==1.23.5  # Изменено на совместимую версию
pandas==1.5.3  # Изменено на совместимую версию
openpyxl==3.1.2
gunicorn==21.2.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
requests==2.31.0
# flower==1.2.0